feat(All): 비즈니스 로직 중복 코드 제거
