+++ README.md
@@ -83,5 +83,5 @@ erDiagram
 
 ### 3. 리팩터링 - 중복 코드 제거
 
-- [ ] 비즈니스 로직 중복 코드 제거
+- [x] 비즈니스 로직 중복 코드 제거
 - [x] 기존 자동차 미션 코드 전체적으로 리팩토링
