+++ src/main/java/racingcar/ConsoleApplication.java
@@ -1,7 +1,10 @@
 package racingcar;
 
 import racingcar.controller.ConsoleController;
-import racingcar.domain.RandomNumberGenerator;
+import racingcar.dao.ConsoleGameDao;
+import racingcar.dao.ConsoleParticipatesDao;
+import racingcar.dao.ConsolePlayerDao;
+import racingcar.service.RacingCarService;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -9,12 +12,10 @@ public class ConsoleApplication {
 
     public static void main(String[] args) {
         ConsoleController consoleController = new ConsoleController(
+                new RacingCarService(new ConsoleGameDao(), new ConsolePlayerDao(), new ConsoleParticipatesDao()),
                 new InputView(),
-                new OutputView(),
-                new RandomNumberGenerator()
+                new OutputView()
         );
-        consoleController.newCarNames();
-        consoleController.newGameRound();
-//        consoleController.play();
+        consoleController.run();
     }
 }
