+++ src/main/java/racingcar/WebApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RacingCarApplication {
+public class WebApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(RacingCarApplication.class, args);
+        SpringApplication.run(WebApplication.class, args);
     }
 
 }
