+++ src/main/java/racingcar/dao/ConsoleGameDao.java
@@ -1,0 +1,17 @@
+package racingcar.dao;
+
+import java.util.List;
+import racingcar.entity.GameEntity;
+
+public class ConsoleGameDao implements GameDao {
+
+    @Override
+    public Long save(final int trialCount) {
+        return null;
+    }
+
+    @Override
+    public List<GameEntity> findAll() {
+        return null;
+    }
+}
