+++ src/main/java/racingcar/dao/ParticipatesDao.java
@@ -1,41 +1,12 @@
 package racingcar.dao;
 
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
 import racingcar.dto.ParticipateDto;
 import racingcar.entity.ParticipatesEntity;
 
-@Repository
-public class ParticipatesDao {
+public interface ParticipatesDao {
 
-    private final JdbcTemplate jdbcTemplate;
+    void save(final ParticipateDto participateDto);
 
-    private final RowMapper<ParticipatesEntity> rowMapper = (resultSet, rowNum) -> new ParticipatesEntity(
-            resultSet.getLong("game_id"),
-            resultSet.getLong("player_id"),
-            resultSet.getInt("position"),
-            resultSet.getBoolean("is_winner"));
-
-    @Autowired
-    public ParticipatesDao(final JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public void save(final ParticipateDto participateDto) {
-        final String sql = "INSERT INTO PARTICIPATES(game_id, player_id, position, is_winner) VALUES(?, ?, ?, ?) ";
-        jdbcTemplate.update(sql,
-                participateDto.getGameId(),
-                participateDto.getPlayerId(),
-                participateDto.getPosition(),
-                participateDto.getIsWinner()
-        );
-    }
-
-    public List<ParticipatesEntity> findByGameId(final Long gameId) {
-        final String sql = "SELECT * FROM PARTICIPATES WHERE game_id = ?";
-        return jdbcTemplate.query(sql, rowMapper, gameId);
-    }
+    List<ParticipatesEntity> findByGameId(final Long gameId);
 }
